# openapi.yaml
{
  "components": {
    "schemas": {
...
      "AuthTokensResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "AuthTokensResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensResponse"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "tokens"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest",
        "type": "object"
      },
      "ProblemDetails": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "org_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Id"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "RegisterResponse": {
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "RegisterResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Auth Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Login",
        "tags": [
          "auth",
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth",
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth",
          "auth"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "operationId": "register_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Invalid credentials",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "https://example.com/problems/unauthorized"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Access denied",
                  "instance": "/v1/admin/users/roles",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 403,
                  "title": "Access Denied",
                  "type": "https://example.com/problems/access-denied"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User not found",
                  "instance": "/v1/admin/users/123e4567-e89b-12d3-a456-426614174000",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 404,
                  "title": "Not Found",
                  "type": "https://example.com/problems/not-found"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "User already exists",
                  "instance": "/v1/auth/register",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 409,
                  "title": "Conflict",
                  "type": "https://example.com/problems/conflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "example": {
                  "detail": "Rate limit exceeded: auth_login",
                  "instance": "/v1/auth/login",
                  "request_id": "4b1c5f8c-7f2b-4e7f-8f7a-6b76a0e3f6a1",
                  "status": 429,
                  "title": "Too Many Requests",
                  "type": "https://example.com/problems/too-many-requests"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Register",
        "tags": [
          "auth",
          "auth"
        ]
      }
    }
  }
}